Improve documentation with examples
- Improve README.md documentation with information on how to specify
different internal & external ports
- Update compose.yaml with an inline `upnp.json` config and some example
services
- Merge updated compose.yaml into README.md

I had to read the code to figure out which port was supposed to be the
internal port and which one was the external in the config. (It wasn't
        even immediately obvious to me that this supported using separate
        internal and external ports, but surely that one's on me), so I figured
improving the docs might help other people as well.

Additionally, I thought that the Compose examples could benefit from
adding `configs:` elements to it. The Compose `configs:` section is
pretty awesome for services like this that require a simple config file
to be mounted because you can define everything you need for
upnp-service within the compose.yaml itself. (It especially makes sense
since users of this will probably be using `ports:` sections within the
same yaml file!)